refactor: move collections logic outside book providers
