--- src/app/providers/local/actions/getAuthorBooks.ts
@@ -1,6 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import type { Book } from '@/types/Book'
-
-export async function getAuthorBooks(_authorId: string): Promise<Book[]> {
-  throw Error('NOT IMPLEMENTED')
-}
