--- src/app/providers/local/actions/getAuthorDetails.ts
@@ -1,6 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import type { Author } from '@/types/Author'
-
-export function getAuthorDetails(_authorId: string): Author {
-  throw Error('NOT IMPLEMENTED')
-}
