--- src/app/providers/local/actions/getBookCovers.ts
@@ -1,4 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-export function getBookCovers(bookId: string): string[] {
-  throw Error('NOT IMPLEMENTED')
-}
