--- src/app/providers/local/actions/getBooksSimilar.ts
@@ -1,6 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import type { Book } from '@/types/Book'
-
-export async function getBooksSimilar(bookId: string): Promise<Book[]> {
-  throw Error('NOT IMPLEMENTED')
-}
