--- src/app/providers/local/actions/getCategoryBooks.ts
@@ -1,6 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import type { Book } from '@/types/Book'
-
-export async function getCategoryBooks(categoryId: string): Promise<Book[]> {
-  throw Error('NOT IMPLEMENTED')
-}
