--- src/app/providers/local/actions/searchBooks.ts
@@ -1,7 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import type { Book } from '@/types/Book'
-import type { BookSearchPayload } from '@/types/BookSearchPayload'
-
-export async function searchBooks(payload: BookSearchPayload): Promise<Book[]> {
-  throw Error('NOT IMPLEMENTED')
-}
