--- src/app/providers/local/index.ts
@@ -1,2 +1,0 @@
-export * from './hooks/useLocalBooksProvider'
-export * from './types'
