--- src/app/providers/local/types.ts
@@ -1,1 +1,0 @@
-export {}
