+++ src/types/BookProvider.ts
@@ -1,8 +1,6 @@
 import type { Author } from '@/types/Author'
 import type { Book } from '@/types/Book'
 import type { BookSearchPayload } from '@/types/BookSearchPayload'
-import type { Collection } from '@/types/Collection'
-import type { CollectionType } from '@/types/CollectionType'
 
 export type BookProvider = {
   searchBooks: (payload: BookSearchPayload) => Async<Book[]>
@@ -12,6 +10,4 @@ export type BookProvider = {
   getBookDetails: (bookId: string) => Async<Book>
   getBooksSimilar: (bookId: string) => Async<Book[]>
   getBookCovers: (bookId: string) => Async<string[]>
-  getCollections: () => Async<Collection[]>
-  getCollectionBooks: (collectionId: CollectionType, userId: string) => Async<Book[]>
 }
